+++ pytket/docs/changelog.rst
@@ -13,6 +13,7 @@ Fixes:
 
 * Fix `symbol_substitution` not preserving opgroups.
 * Remove hardware inefficient circuit construction in `_tk1_to_rzsx`
+* Support converting conditional `RangePredicate`s to QASM.
 
 1.34.0 (October 2024)
 ---------------------